Fix: Cannot find module '${HOME}/host.js'
It appears that Firefox starts the native client from `$HOME` as the working directory.
That leads to the script resolving the relative path host.js as `$HOME/host.js`.

This fix prefixes the `host.js` filename with the directory of the run.sh file.
Now it works regardless of the working directory and regardless of the directory in which
`run.sh` and the JavaScript files are located (as long as they are in the same location).

Fix: andy-portmen/native-client#118